% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{nanoarrow}
\alias{nanoarrow}
\alias{as_nanoarrow_array.HArray}
\alias{infer_nanoarrow_ptype.HArray}
\alias{as_arrow_array.HArray}
\alias{as_nanoarrow_array.HMatrix}
\alias{infer_nanoarrow_ptype.HMatrix}
\alias{as_arrow_array.HMatrix}
\title{HArray and HMatrix to nanoarrow and arrow}
\usage{
as_nanoarrow_array.HArray(x)

infer_nanoarrow_ptype.HArray(x)

as_arrow_array.HArray(x)

as_nanoarrow_array.HMatrix(x)

infer_nanoarrow_ptype.HMatrix(x)

as_arrow_array.HMatrix(x)
}
\arguments{
\item{x}{an HArray.}
}
\value{
A nanoarrow array.

An R vector of zero size describing the target into which the array should be materialized.

An arrow array.

A nanoarrow array.

An R vector of zero size describing the target into which the array should be materialized.

An arrow array.
}
\description{
Conversion via native apache arrow array. Requires ´nanoarrow´ package.
}
\details{
The following functions enable conversion to \code{nanoarrow} and \code{arrow}.
}
\examples{
library(nanoarrow)
harray = HArray$new_from_values(c(1,2,3), HDataType$float32())
nanoarrow_array = as_nanoarrow_array(harray)
v = as.vector(nanoarrow_array)
print(v)

infer_nanoarrow_ptype(harray)

arrow::as_arrow_array(harray)
library(nanoarrow)
values = matrix(c(1,2,3,4), ncol = 2)
hmatrix = HMatrix$new_from_values(values, HDataType$float32())
nanoarrow_array = as_nanoarrow_array(hmatrix)
v = as.vector(nanoarrow_array)
print(v)
infer_nanoarrow_ptype(hmatrix)

arrow::as_arrow_array(hmatrix)
}
\keyword{nanoarrow_interface.}
