% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structs.R
\name{print.HMetadataType}
\alias{print.HMetadataType}
\title{print a HMetadataType}
\usage{
\method{print}{HMetadataType}(x)
}
\arguments{
\item{x}{HMetadataType}
}
\value{
self
}
\description{
print a HMetadataType
}
