% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structs.R
\name{print.HResamplerType}
\alias{print.HResamplerType}
\title{print a HResamplerType}
\usage{
\method{print}{HResamplerType}(x)
}
\arguments{
\item{x}{HResamplerType}
}
\value{
self
}
\description{
print a HResamplerType
}
